+++ R/dsl.R
@@ -96,10 +96,12 @@ add_code_step <- function(stage, call = NULL, prepare_call = NULL) {
 #' @rdname DSL
 #' @export
 #' @importFrom magrittr %>%
-add_package_checks <- function(warnings_are_errors = TRUE,
-                               notes_are_errors = FALSE,
+add_package_checks <- function(...,
+                               warnings_are_errors = NULL,
+                               notes_are_errors = NULL,
                                args = c("--no-manual", "--as-cran"),
-                               build_args = "--force") {
+                               build_args = "--force", error_on = "warning",
+                               repos = getOption("repos"), timeout = Inf) {
   #' @description
   #' 1. A [step_rcmdcheck()] in the `"script"` stage, using the
   #'    `warnings_are_errors`, `notes_are_errors`, `args`, and
@@ -110,7 +112,10 @@ add_package_checks <- function(warnings_are_errors = TRUE,
         warnings_are_errors = warnings_are_errors,
         notes_are_errors = notes_are_errors,
         args = args,
-        build_args = build_args
+        build_args = build_args,
+        error_on = error_on,
+        repos = repos,
+        timeout = timeout
       )
     )