+++ NEWS.md
@@ -1,7 +1,17 @@
+# tic 0.2.13.9011
+
+## step_rcmdcheck()
+
+- deprecate `warnings_are_errors` and `notes_are_errors` in favor of the new `error_on` argument
+- add args `timeout` and `repos`
+- call `rcmdcheck()` internally with `error_on = "never"` so that we can trigger the message on found warnings and notes
+- remote outdated doc about `step_rcmdcheck()` using a dedicated lib for the check
+
 # tic 0.2.13.9010
 
 - No longer using a separate library for package checks, because it causes a lot of problems with various steps which are not aware of this (#86, #88).
-
+- Packages coming with the R-installation are not updated anymore when preparing `step_rcmdcheck()`. 
+  See `?step_rcmdcheck()` for detailed info. (#103)
 
 # tic 0.2.13.9009